requirements.txt
django==3.2.2
djangorestframework==3.12.4
python-dotenv==0.19.0
django-rest-swagger
pytest==6.2.5
